cqlsh: support --file parameter
use the -f option with an input filename.

this is more than just "cqlsh < myfile.cql": when encountering an error
while reading from an input file, cqlsh will print out the filename and
line number of the error, for easier debugging.

this commit also involves replacing the implementation of
cmd.Cmd.cmdloop(), since it provided no way to tell between the text
"eof" appearing on a line by itself and an actual eof occurring on
input.